+++ src/screens/organizer/index.ts
@@ -2,3 +2,4 @@ export * from './EventsScreen';
 export * from './LogisticsScreen';
 export * from './MeetupsScreen';
 export * from './VerifyCosplayersScreen';
+export * from './VerifyStaffScreen';